faq-video-poc/requirements.txt
# Core dependencies
chromadb>=0.4.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
pandas>=1.5.0
numpy>=1.21.0
click>=8.0.0

# WebRTC for real-time video streaming
aiortc>=1.6.0
av>=10.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and utilities
loguru>=0.6.0
tqdm>=4.64.0

# Optional: For advanced NLP processing
transformers>=4.21.0
torch>=1.12.0

# Optional: For web interface
fastapi>=0.100.0
uvicorn>=0.23.0
